ist-9 Refactor for better code organization/separation
