+++ backend/pyproject.toml
@@ -7,7 +7,8 @@ readme = "README.md"
 packages = [
     {include = "api"},
     {include = "database"},
-    {include = "internet_speed_tracker"}
+    {include = "speedtracker"},
+    {include = "shared"}
 ]
 
 [tool.poetry.dependencies]
